apps/backend/requirements.txt
Django<5
djangorestframework
drf-yasg
django-cors-headers
pydantic
psycopg2-binary
markdown
bleach
python-slugify

